advanced

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -9,7 +9,7 @@
 int main(void)
 {
 	int i;
-	long long int k;
+	long int k;
 
 	i = 2;
 	k = 612852475143;
